move resolve_path to Session inherent method
