+++ compiler/rustc_session/src/diagnostics.rs
@@ -721,3 +721,11 @@ pub(crate) struct NativeTargetCpuNotAllowed<'a> {
     pub(crate) target_triple: &'a TargetTuple,
     pub(crate) need_explicit_cpu: bool,
 }
+
+#[derive(Diagnostic)]
+#[diag("cannot resolve relative path in non-file source `{$path}`")]
+pub(crate) struct ResolveRelativePath {
+    #[primary_span]
+    pub span: Span,
+    pub path: String,
+}
